Simple Python Database with MVC Design
This project is my attempt to create a simple database that uses Ktinker for a GUI, that can take in string data from the user and store it in a JSON file that can also be displayed using the GUI to the user. The project uses MVC (Main - View - Controller) design pattern in order to accomplish this!